\name{proc}
\alias{proc}
\docType{data}
\title{Production data}
\description{Data on fictitious production processes.}
\usage{data(proc)}
\format{
  A data frame with a total of 60 observations on the following 3 variables:
  \describe{
    \item{\code{output}}{The amount of output produced.}
    \item{\code{capital}}{The amount of capital utilized.}
    \item{\code{labour}}{The amount of labour employed.}
  }
}
