\name{prod}
\alias{prod}
\docType{data}
\title{Simulated production data}
\description{Simulated data for production processes with one output variable and two input variables.}
\usage{data(prod)}
\format{
  A data.frame with a total of 60 observations on the following 3 variables:
  \describe{
    \item{\code{output}}{The amount of output produced.}
    \item{\code{capital}}{The amount of capital utilized.}
    \item{\code{labour}}{The amount of labour employed.}
  }
}
\references{
T. Coelli (1996). A Guide to FRONTIER Version 4.1: A Computer Program for Stochastic Frontier Production and Cost Function Estimation. CEPA Working Paper 96/08, University of New England.
}
