\name{production}
\alias{production}
\docType{data}
\title{Simulated production data}
\description{Simulated data for production processes with one output variable and two input variables.}
\usage{data(production)}
\format{
  A data.frame with a total of 56 observations on the following 3 variables:
  \describe{
    \item{\code{output}}{The amount of output produced.}
    \item{\code{capital}}{The amount of capital utilized.}
    \item{\code{labour}}{The amount of labour employed.}
  }
}
