\name{rice}
\alias{rice}
\docType{data}
\title{Rice production data}
\description{Data on several fictitious rice production processes.}
\usage{data(rice)}
\format{
  A data frame with a total of 100 observations on the following 5 variables:
  \describe{
    \item{\code{prod}}{The amount of rice produced (tonnes).}
    \item{\code{area}}{The amount of area planted (hectares).}
    \item{\code{labour}}{The amount of labour employed (man-days).}
    \item{\code{fertil}}{The amount of fertilizer used (kilograms).}
    \item{\code{machinery}}{The amount of machinery utilized (index, firm 41=100).}
  }
}
