\name{rice}
\alias{rice}
\docType{data}
\title{Rice production data}
\description{Data on fictitious processes of rice production.}
\usage{data(rice)}
\format{
  A data.frame with a total of 100 observations on the following 5 variables:
  \describe{
    \item{\code{rice}}{The amount of rice produced (tonnes).}
    \item{\code{land}}{The amount of land planted (hectares).}
    \item{\code{labour}}{The amount of labour employed (man-days).}
    \item{\code{fertil}}{The amount of fertilizer used (kilograms).}
    \item{\code{machinery}}{The amount of machinery utilized (index, firm 41=100).}
  }
}
